requirements.txt
langchain-ollama
langchain
ollama
langchain-community
langchain
langchain-core
langchain-text-splitters
torch
transformers
sentence-transformers
scikit-learn
faiss-cpu
openai
tiktoken
numpy